refactor: 댓글 삭제할 때 주인이 아닌 경우 예외 발생하는 기능 구현

+++ src/main/java/com/example/dailychallenge/service/comment/CommentService.java
@@ -92,8 +92,10 @@ public class CommentService {
         return comment;
     }
 
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, User user) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(EntityNotFoundException::new);
+        validateOwner(user, comment);
+
         commentRepository.delete(comment);
     }
